+++ rog-slash/src/usb.rs
@@ -37,8 +37,10 @@ pub fn get_slash_type() -> SlashType {
     let dmi = DMIID::new()
         .map_err(|_| SlashError::NoDevice)
         .unwrap_or_default();
-    let board_name = dmi.board_name;
-    if board_name.contains("GA403W") {
+    let board_name = dmi.board_name.to_uppercase();
+    if board_name.contains("G614F") {
+        SlashType::G614F
+    } else if board_name.contains("GA403W") {
         SlashType::GA403W
     } else if board_name.contains("GA403") {
         SlashType::GA403
@@ -58,6 +60,7 @@ pub const fn report_id(slash_type: SlashType) -> u8 {
         SlashType::GA403W => REPORT_ID_19B6,
         SlashType::GA403 => REPORT_ID_193B,
         SlashType::GA605 => REPORT_ID_19B6,
+        SlashType::G614F => REPORT_ID_19B6,
         SlashType::GU605 => REPORT_ID_193B,
         SlashType::GU605C => REPORT_ID_19B6,
         SlashType::Unsupported => REPORT_ID_19B6,